Added github template for Issues and Pull Requests
I was working together with @make-it-git today and he opened a PR without
a description, and var.ci bot was repeatedly complaining and closing
the issue as invalid.

This wasted quite a bit of time, so I think the solution is to introduce
templates for opening PRs and for reporting issues.

This will help ensure consistent PRs and Issues, and will help keep any
new developers joining the project on the same page.