Fix for broken test
The issue was that the `testBadImport()` class' second parameter
needs to be the expected value to fail on.

ID is a valid import, and if you look at this defined method you will
see that it makes a single object, with a single property ID, and then
attempts to load it into the model.

This means that the import fails on the first missing property, which,
if you look in the mock, is date.

If you look carefully at the error message you'll see that that failure
is a string assertion that deceptively looks like a single error, but is
in fact a comparison of 2 strings.

I'll link a screenshot.